simple settlement building

+++ Siedler4Bot/src/utils/MathUtils.hpp
@@ -4,5 +4,5 @@
 
 inline unsigned short CalcDistance(unsigned short x1, unsigned short y1, unsigned short x2, unsigned short y2) noexcept
 {
-    return std::sqrt(std::pow(x2 - x1, 2) + std::pow(y2 - y1, 2));
+    return static_cast<unsigned short>(std::sqrt(std::pow(x2 - x1, 2) + std::pow(y2 - y1, 2)));
 }
